fix for nominette
